pyproject.toml
[tool.poetry]
name = "repl_python3_Normal-Bot-4"
version = "0.1.0"
description = ""
authors = ["Kavoya <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.2"
discord = "^1.7.3"
Flask = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
